LayoutTests/storage/indexeddb/objectstore-removeobjectstore-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
Test IndexedDB's webkitIDBObjectStore.deleteObjectStore().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.open('objectstore-removeobjectstore')
db = event.target.result
db.setVersion('new version')
trans = event.target.result
PASS trans is non-null.
Deleted all object stores.
store = db.createObjectStore('storeName', null)
store.createIndex('indexName', '')
PASS store.indexNames.contains('indexName') is true
store.add('value', 'key')
db.transaction(['storeName'])
store = transaction.objectStore('storeName')
store.get('key')
PASS event.target.result is "value"
db.setVersion('new version')
trans = event.target.result
PASS trans is non-null.
db.deleteObjectStore('storeName')
db.createObjectStore('storeName', null)
db.transaction(['storeName'])
store = transaction.objectStore('storeName')
store.get('key')
PASS event.target.result is undefined
PASS event.target.source.indexNames.contains('indexName') is false
PASS successfullyParsed is true

TEST COMPLETE
